Api permissions and adding model and auth classes

+++ models/admin.js
@@ -1,19 +1,21 @@
 
-    class Admin {
+    const Auth = require('../modules/Auth')
+
+    class Admin extends Auth {
 
         constructor(data){
 
-            this.data = data
+            super(data)
 
             this.settings = {
                 collection: 'admin',
                 fields: [
-                    {name:'avatar',input_type:'image',placeholder:'Profile Picture', type:'image'},
-                    {name:'full_name',input_type:'text',placeholder:'Full name', type:'string'},
-                    {name:'email',input_type:'email',placeholder:'Email address', type:'string'},
-                    {name:'tel',input_type:'text',placeholder:'Telephone Number', type:'string'},
-                    {name:'activated',input_type:'select',options:[{text:'Yes',value:true},{text:'No',value:false}], type:'boolean'},
-                    {name:'blocked',input_type:'select',options:[{text:'Yes',value:true},{text:'No',value:false}], type:'boolean'}
+                    {name:'avatar',input_type:'image',placeholder:'Profile Picture', type:'image', required:false},
+                    {name:'full_name',input_type:'text',placeholder:'Full name', type:'name', required:true},
+                    {name:'email',input_type:'email',placeholder:'Email address', type:'email', required:true},
+                    {name:'tel',input_type:'text',placeholder:'Telephone Number', type:'tel', required:false},
+                    {name:'activated',input_type:'select',options:[{text:'Yes',value:true},{text:'No',value:false}], type:'boolean', required:false},
+                    {name:'blocked',input_type:'select',options:[{text:'Yes',value:true},{text:'No',value:false}], type:'boolean', required:false}
                 ]
             }
 
@@ -45,91 +47,6 @@
 
         }
 
-        find(key) {
-
-            if (key){
-
-                let field
-
-                if (typeof key == 'object'){
-
-                    if (key._key){
-
-                        key = key._key
-                        field = '_key'
-
-                    } else if (key.email){
-
-                        key = key.email
-                        field = 'email'
-
-                    } else if (key.password_reset){
-
-                        key = key.password_reset
-                        field = 'password_reset'
-
-                    }
-
-                } else {
-
-                    if (typeof key == 'string' && key.match(/^[0-9]*$/)){
-
-                        if (key._key){
-                            key = key._key
-                        }
-
-                        field = '_key'
-
-                    } else if (typeof key == 'string' && key.match(/@/)){
-
-                        if (key.email){
-                            key = key.email
-                        }
-
-                        field = 'email'
-
-                    } else if (key){
-
-                        field = 'password_reset'
-
-                    }
-
-                }
-
-                this.data = db.read(this.settings.collection)
-                                    .where([field+' == '+key])
-                                    .first(['password','password_reset'])
-
-                if (this.data){
-                    this.data.guard = this.settings.collection
-                    return this
-                } else {
-                    this.error = 'Not found'
-                    return this
-                }
-
-            } else {
-                this.error = 'No key provided'
-                return this
-            }
-
-        }
-
-        all() {
-
-            this.data = db.read(this.settings.collection).orderBy('full_name','asc').omit(['password','password_reset']).get()
-            return this.data
-
-        }
-
-        get(){
-            return this.data
-        }
-
-        first(){
-            return this.data[0]
-        }
-
         search(search) {
 
             if (search.str.length < 3){
@@ -159,42 +76,7 @@
                 return this
             }
 
-            if (this.data.tel){
-                this.data.tel = this.data.tel.replace(/\s/g,'')
-            }
-
-            if (typeof this.data.full_name == 'string'){
-                let name = this.data.full_name.split(" ")
-                this.data.name = {}
-                this.data.name.first = view.functions.capitalise(name[0])
-                this.data.name.last = view.functions.capitalise(name[name.length-1])
-                this.data.full_name = view.functions.capitalise(this.data.full_name)
-            } else if (this.data.name) {
-                this.data.full_name = view.functions.capitalise(this.data.name.first+' '+this.data.name.last)
-            }
-
-            // if (this.data.name && this.data.name.first && this.data.name.last){
-            //     images.saveAll(data,this.data.name.first+'-'+this.data.name.last,'avatars').then((new_data)=>{
-            //         data = new_data
-            //     })
-            // }
-
-            if (this.data.password && this.data.password_conf && this.data.password == this.data.password_conf){
-
-                this.data.password = db.hash(this.data.password)
-                delete this.data.password_conf
-
-            } else if (this.data.password && this.data.password_conf && this.data.password == this.data.password_conf){
-
-                return 'Passwords do not match'
-
-            } else {
-
-                delete this.data.password
-
-            }
-
-        //    this.data = db.parseDataTypes(this.settings.collection, this.data)
+            await this.validate()
 
             if (this.data._id){
 
@@ -214,18 +96,27 @@
 
             }
 
-            return this.data
+            return this
 
         }
 
-        delete(){
+        async delete(){
 
-            this.data = db.read(this.settings.collection).where(['_key == '+this.data._key]).delete()
-            if (this.data.length > 0){
-                this.error = 'Not deleted'
+            let admin_count = await db.count(this.settings.collection)
+
+            if (admin_count > 1){
+                this.data = db.read(this.settings.collection).where(['_key == '+this.data._key]).delete()
+                if (this.data.length > 0){
+                    this.error = 'Not deleted'
+                    return this
+                } else {
+                    return this.data
+                }
             } else {
-                return this.data
+                this.error = 'At least 1 admin account must be registered'
+                return this
             }
+
 
         }
 
@@ -251,6 +142,7 @@
             }
 
         }
+
 
     }
